+++ contracts/interfaces/ISecurityToken.sol
@@ -40,7 +40,7 @@ contract ISecurityToken is IST20, Ownable {
      * @param _moduleType is which type of module we are trying to remove
      * @param _name is the name of the module within the chosen type
      */
-    function getModuleByName(uint8 _moduleType, bytes32 _name) public view returns (bytes32, address);
+    function getModuleByName(uint8 _moduleType, bytes32 _name) public view returns (bytes32[], address[]);
 
     /**
      * @notice Queries totalSupply as of a defined checkpoint